Reorganisation of the projetc structur and set some visual changes
