+++ client/package.json
@@ -8,11 +8,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "fuse": "^0.12.1",
     "fuse.js": "^7.0.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^6.26.2",
-
     "react-slick": "^0.30.2",
     "slick-carousel": "^1.8.1"
   },
