+++ client/src/App.tsx
@@ -1,11 +1,12 @@
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import "./CssComponents/Crrousels.css";
-import Header from "./components/header";
+import Header from "./components/Header/header";
+import HomePage from "./components/HomePage/HomePage";
+import "./CssComponents/CarousselsHomePage.css";
+
 export default function App() {
   return (
     <>
       <Header />
+      <HomePage />
     </>
   );
 }
