+++ client/src/CssComponents/CarousselsHomePage.css
@@ -76,7 +76,7 @@ body {
 .slick-slide img {
   width: 100%;
   height: auto;
-  border-radius: 10px;
+  border-radius: 15px;
   transition: 0.7s ease-in-out;
 }
 
@@ -96,7 +96,7 @@ body {
 }
 
 .slick-prev {
-  margin-right: 2rem;
+  left: -43px;
 }
 
 .slick-next:before {
