+++ client/src/components/AnimsList/animeCard.css
@@ -1,0 +1,23 @@
+@import url("https://fonts.googleapis.com/css2?family=Lato:ital,wght@0,100;0,300;0,400;0,700;0,900;1,100;1,300;1,400;1,700;1,900&family=Sour+Gummy:ital,wght@0,100..900;1,100..900&display=swap");
+
+.figureAnime {
+  position: relative;
+  border-radius: 15px;
+  font-size: larger;
+  font-family: "Lato", sans-serif;
+  display: flex;
+}
+.figcaptionAnime {
+  position: absolute;
+  background-color: rgba(0, 0, 0, 1);
+  color: whitesmoke;
+  opacity: 0;
+  border-radius: 15px;
+}
+
+.figcaptionAnime:hover {
+  opacity: 0.7;
+  width: 90%;
+  height: 100%;
+  transition: opacity 0.5s ease;
+}
