+++ client/src/components/AnimsList/animeCard.tsx
