+++ client/src/components/AnimsList/animeList.css
@@ -1,0 +1,20 @@
+.animeContainer {
+  width: 95%;
+  margin-left: 3rem;
+}
+.animePicture {
+  width: 90%;
+  border-radius: 15px;
+}
+
+h2 {
+  font-size: 24px;
+  margin-bottom: 2rem;
+  margin-top: 2rem;
+}
+
+.anime-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr;
+  align-items: center;
+}
