+++ client/src/components/Header/header.css
@@ -1,0 +1,32 @@
+@import url("https://fonts.googleapis.com/css2?family=Jersey+10&family=Sour+Gummy:ital,wght@0,100..900;1,100..900&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Host+Grotesk:ital,wght@0,300..800;1,300..800&family=Jersey+10&family=Sour+Gummy:ital,wght@0,100..900;1,100..900&display=swap");
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+}
+
+header {
+  height: 18vh;
+  /*background-color: #1b263b;*/
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-around;
+}
+
+nav {
+  display: flex;
+  gap: 2rem;
+}
+
+a {
+  font-family: "Sour Grummy", sans-serif;
+  text-decoration: none;
+  color: whitesmoke;
+}
+
+header img {
+  width: 130px;
+  height: 130px;
+}
