+++ client/src/components/Header/header.tsx
@@ -1,13 +1,11 @@
 import { Link, Outlet } from "react-router-dom";
-import logo from "../../public/boite_de_popcorn_0.png";
 import "./header.css";
-
 export default function Header() {
   return (
     <>
       <header>
         <Link to="/">
-          <img src={logo} width={25} height={44} alt="logo" />
+          <img src="./public/pop-corn_Box.png" alt="logo Pop Corn Box" />
         </Link>
 
         <nav>
@@ -18,7 +16,6 @@ export default function Header() {
         </nav>
 
         <input type="text" placeholder="Rechercher..." />
-        <Link to="/MesPréferences">Préferences</Link>
         <Link to="/WatchList">WatchList</Link>
       </header>
       <Outlet />
