+++ client/src/components/MovieList/style.css
@@ -18,6 +18,8 @@ body {
 h1,
 h2,
 h3 {
+  margin-bottom: 2rem;
+  margin-top: 2rem;
   font-family: "Sour Gummy", sans-serif;
 }
 
@@ -30,6 +32,7 @@ p {
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
   justify-self: center;
+  margin-left: 2rem;
 }
 
 .movieList {
