--- client/src/components/animeCard.css
@@ -1,23 +1,0 @@
-.figureAnime {
-  position: relative;
-  margin: 0;
-  border-radius: 1.8rem;
-  overflow: hidden;
-}
-.figcaptionAnime {
-  display: none;
-  flex-direction: column;
-  justify-content: center;
-  left: 50%;
-  top: 50%;
-  transform: translate(-50%, -50%);
-  position: absolute;
-  background-color: rgba(0, 0, 0, 0.7);
-  width: 100%;
-  height: 100%;
-  color: blanchedalmond;
-}
-
-figure:hover figcaption {
-  display: flex;
-}
