--- client/src/components/animeList.css
@@ -1,20 +1,0 @@
-.animeContainer {
-  margin: auto;
-  text-align: center;
-  width: 80%;
-}
-.animePicture {
-  width: 100%;
-}
-
-h1 {
-  font-size: 24px;
-  margin-bottom: 1rem;
-  margin-top: 1rem;
-}
-
-.anime-grid {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1rem;
-}
