--- client/src/components/header.css
@@ -1,37 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-}
-
-header {
-  height: 18vh;
-  background-color: #1b263b;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-around;
-}
-
-nav {
-  display: flex;
-  gap: 2rem;
-}
-
-a {
-  color: aliceblue;
-  text-decoration: none;
-}
-ul {
-  display: flex;
-  justify-content: center;
-}
-
-header img {
-  width: 4rem;
-  margin-left: 1.4rem;
-}
-
-li {
-  padding-right: .2rem;
-  margin: .5rem;
-}
